requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Login==0.6.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
twilio==8.1.0
passlib==1.7.4
itsdangerous==2.1.2
Werkzeug==2.2.3
email-validator==2.0.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Pillow==9.5.0
Flask-JWT-Extended==4.5.2
plotly==5.14.1
pandas==2.0.0
wtforms==3.0.1
Flask-Cors==3.0.10
requests==2.31.0 